Changes to question adding to sessions
