--- client/css/components/AddQuestionModal.scss
@@ -1,5 +1,0 @@
-
-.ql-modal-addquestion {
-
-  
-}
