+++ client/css/components/QuestionEditItem.scss
@@ -10,7 +10,6 @@
   .answer-no-wysiwyg {
     font-size: 130%;
     font-weight: bold;
-    background: #eee;
     padding: 10px;
     text-align: center;
   }
@@ -35,7 +34,6 @@
   }
 
   .editor-wrapper {
-    background: #eee;
     
     .rdw-option-wrapper {
       height: 30px;
