+++ client/css/components/QuestionSearch.scss
@@ -1,0 +1,5 @@
+
+.ql-question-sidebar {
+
+  
+}
