+++ imports/ui/QuestionEditItem.jsx
@@ -21,7 +21,7 @@ import { MC_ORDER, TF_ORDER, QUESTION_TYPE, QUESTION_TYPE_STRINGS, EDITOR_OPTION
 
 export const DEFAULT_STATE = {
   question: '',
-  type: QUESTION_TYPE.MC, // QUESTION_TYPE.MC, QUESTION_TYPE.TF, QUESTION_TYPE.SA
+  type: -1, // QUESTION_TYPE.MC, QUESTION_TYPE.TF, QUESTION_TYPE.SA
   content: null,
   answers: [], // { correct: false, answer: 'A', content: editor content }
   submittedBy: '',
@@ -58,7 +58,7 @@ export class QuestionEditItem extends ControlledForm {
       this.state = _.extend({}, DEFAULT_STATE)
 
       // bold placehodler text for question editor
-      const initialQuestionState = ContentState.createFromBlockArray(convertFromHTML('<h2>&nbsp; ?</h2>').contentBlocks)
+      const initialQuestionState = ContentState.createFromBlockArray(convertFromHTML('<h2>Question?</h2>').contentBlocks)
       this.state.content = EditorState.createWithContent(initialQuestionState)
     }
     // set draft-js editor toolbar options
@@ -225,32 +225,7 @@ export class QuestionEditItem extends ControlledForm {
       return
     }
 
-    // convert draft-js objects into JSON
-    const serialized = DraftHelper.toJson(question.content)
-    question.content = serialized.json
-    question.question = serialized.plainText
-
-    question.answers = []
-    this.state.answers.forEach((a) => {
-      let copy = _.extend({}, a)
-      if (copy.wysiwyg) {
-        const serialized = DraftHelper.toJson(copy.content)
-        copy.content = serialized.json
-        copy.plainText = serialized.plainText
-      } else if (question.type === QUESTION_TYPE.TF) {
-        copy.content = a.answer
-        copy.plainText = a.answer
-      }
-
-      question.answers.push(copy)
-    })
-
-    question.tags = []
-    this.state.tags.forEach((t) => {
-      question.tags.push(_.extend({}, t))
-    })
-
-    question.courseId = this.props.courseId
+    // TODO set session id
 
     Meteor.call('questions.insert', question, (error) => {
       if (error) {
@@ -342,7 +317,10 @@ export class QuestionEditItem extends ControlledForm {
     return (
       <div className='ql-question-edit-item'>
         <div className='header'>
-          <h2>{ !this.state._id ? 'Add a Question' : 'Edit Question' }</h2>
+
+          { newEditor(this.state.content, this.onEditorStateChange) }
+
+          {/*
           <select defaultValue={this.state.type} onChange={this.changeType} className='question-type form-control'>
             {
               _(QUESTION_TYPE).keys().map((k) => {
@@ -350,32 +328,49 @@ export class QuestionEditItem extends ControlledForm {
                 return <option key={k} value={val}>{ QUESTION_TYPE_STRINGS[val] }</option>
               })
             }
-          </select>
-          { this.state.type === QUESTION_TYPE.MC
-            ? <button className='btn btn-default' onClick={this.addAnswer}>Add Answer</button>
-            : '' }
+          </select> */}
 
+          <div className='ql-prompt-option'>
+            MC Icon
+            <input type='radio' value={QUESTION_TYPE.MC} name='question-type[]' onChange={this.changeType} />
+          </div>
+
+          <div className='ql-prompt-option'>
+            MS Icon
+            <input type='radio' value={QUESTION_TYPE.MS} name='question-type[]' onChange={this.changeType} />
+          </div>
+
+          <div className='ql-prompt-option'>
+            TF Icon
+            <input type='radio' value={QUESTION_TYPE.TF} name='question-type[]' onChange={this.changeType} />
+          </div>
+
+          <div className='ql-prompt-option'>
+            SA Icon
+            <input type='radio' value={QUESTION_TYPE.SA} name='question-type[]' onChange={this.changeType} />
+          </div>
         </div>
+
+
+        { this.state.type === QUESTION_TYPE.MC || this.state.type === QUESTION_TYPE.MS
+          ? <button className='btn btn-default' onClick={this.addAnswer}>Add Answer</button>
+          : '' }
 
         <form ref='questionForm' className='ql-form-question' onSubmit={this.handleSubmit}>
           <div className='row'>
-            <div className='col-md-8'>{ newEditor(this.state.content, this.onEditorStateChange) }</div>
-            <div className='col-md-4'>
+
+            {/*<div className='col-md-4'>
               <h3>Tags</h3>
               <ReactTags ref='tagInput' tags={this.state.tags}
                 suggestions={this.tagSuggestions}
                 handleDelete={this.deleteTag}
                 handleAddition={this.addTag}
                 handleDrag={this.handleDrag} />
-            </div>
+            </div> */}
           </div>
 
           { editorRows }
 
-          <div className='ql-buttongroup'>
-            <a className='btn btn-default' onClick={this.done}>Cancel</a>
-            <input className='btn btn-default' type='submit' id='submit' />
-          </div>
         </form>
 
       </div>)
