+++ imports/ui/QuestionListItem.jsx
@@ -18,32 +18,34 @@ export class QuestionListItem extends Component {
   constructor (props) {
     super(props)
 
-    const qid = this.props.question._id
+    if (this.props.question) {
+      const qid = this.props.question._id
 
-    const stopClick = (e) => {
-      e.preventDefault()
-      e.stopPropagation()
+      const stopClick = (e) => {
+        e.preventDefault()
+        e.stopPropagation()
+      }
+
+      this.click = this.props.click ? (e) => {
+        stopClick(e)
+        this.props.click(qid)
+      } : noop
+
+      this.remove = this.props.remove ? (e) => {
+        stopClick(e)
+        this.props.remove(qid)
+      } : noop
+
+      this.delete = this.props.delete ? (e) => {
+        stopClick(e)
+        this.props.delete(qid)
+      } : noop
     }
-
-    this.click = this.props.click ? (e) => {
-      stopClick(e)
-      this.props.click(qid)
-    } : noop
-
-    this.remove = this.props.remove ? (e) => {
-      stopClick(e)
-      this.props.remove(qid)
-    } : noop
-
-    this.delete = this.props.delete ? (e) => {
-      stopClick(e)
-      this.props.delete(qid)
-    } : noop
   }
 
   render () {
     // const navigateToSession = () => { Router.go('session', { _id: this.props.session._id }) }
-    const q = this.props.question
+    const q = this.props.question || { question: 'Question?', type: 0 }
     return (
       <li
         className={this.props.click ? 'cursor-pointer' : '' + ' ql-question-list-item'}
@@ -64,7 +66,7 @@ export class QuestionListItem extends Component {
 }
 
 QuestionListItem.propTypes = {
-  question: PropTypes.object.isRequired,
+  question: PropTypes.object,
   click: PropTypes.func,
   remove: PropTypes.func,
   delete: PropTypes.func
